refactor: reactvie 관련 폴더변경
