+++ src/js/reactive/index.js
@@ -1,0 +1,4 @@
+import reactiveState from './reactiveState.js';
+import { registeReactiveRender, reactiveRender } from './reactiveRender.js';
+
+export { reactiveState, registeReactiveRender, reactiveRender };
